examples/requirements.txt
databricks-sql-connector
databricks-sdk
pyspark
python-dotenv
pandas
../dist/beaker-1.0.0-py3-none-any.whl